refactor(pci-instances): isolate add sshkey form from the creation root form
